+++ vllm_ascend/patch/platform/__init__.py
@@ -21,6 +21,7 @@ import vllm_ascend.patch.platform.patch_distributed  # noqa
 import vllm_ascend.patch.platform.patch_kv_cache_utils  # noqa
 import vllm_ascend.patch.platform.patch_mla_prefill_backend  # noqa
 import vllm_ascend.patch.platform.patch_pp_mtp  # noqa
+import vllm_ascend.patch.platform.patch_use_v2_model_runner  # noqa
 from vllm_ascend.utils import is_310p, vllm_version_is
 
 if not is_310p():
