+++ vllm_ascend/patch/worker/patch_v2/patch_use_v2_model_runner.py
@@ -1,20 +1,3 @@
-import vllm.envs as envs
-from vllm.config.vllm import VllmConfig
-
-
-def _patched_use_v2_model_runner(self) -> bool:
-    """Return VLLM_USE_V2_MODEL_RUNNER env directly.
-
-    The upstream use_v2_model_runner gate-keeps the v2 runner with
-    per-model architecture whitelists, Triton availability checks, and
-    feature-support inspections. On Ascend the v2 runner is controlled
-    purely by the VLLM_USE_V2_MODEL_RUNNER environment variable;
-    model-compatibility decisions are deferred to the NPU runner itself.
-    """
-    use_v2 = envs.VLLM_USE_V2_MODEL_RUNNER
-    if use_v2 is not None:
-        return use_v2
-    return False
-
-
-VllmConfig.use_v2_model_runner = property(_patched_use_v2_model_runner)
+# Reuse the platform patch. EngineCore subprocesses only load global/platform
+# patches, while workers also import this compatibility module.
+import vllm_ascend.patch.platform.patch_use_v2_model_runner  # noqa: F401
